fix issue with PUT /book/{id}

+++ src/main/java/com/example/honmon/Controllers/Api/BookController.java
@@ -24,6 +24,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Collections;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -41,9 +42,12 @@ import com.example.honmon.storage.StoredFile;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.apache.commons.io.FilenameUtils;
-import org.apache.commons.io.IOUtils;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -52,11 +56,14 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
+
+import ch.qos.logback.classic.Logger;
 
 @RestController
 @RequestMapping("api/books")
@@ -69,14 +76,24 @@ public class BookController {
     @Autowired
     StorageService<StoredFile> storageService;
 
+    @Autowired
+    private MongoTemplate mongoTemplate;
+
     BookController(BookRepository bookRepository)
     {
         this.bookRepository = bookRepository;
     }
 
     @GetMapping()
-    public List<Book> all()
+    public List<Book> all(
+        @RequestParam(required = false) String key,
+        @RequestParam(required = false) String val
+    )
     {
+        if (key != null && key.equals("title")) {
+            return bookRepository.findBooksByRegexpTitle(val);            
+        }
+        
         return (List<Book>) bookRepository.findAll();
     }
 
@@ -111,15 +128,73 @@ public class BookController {
 
     @GetMapping("{id}")
     public Book getBook(@PathVariable String id) {
-        return bookRepository.findById(id);
+        Book book = bookRepository.findById(id);
+        book.setLastAccess(new Date());
+        bookRepository.save(book);
+        // try {
+            //book.setThumbnail(storageService.load(book.getThumbnailId()));
+            //book.setBook(storageService.load(book.getBookId()));
+        // } catch (IOException e) {
+            // e.printStackTrace();
+        // }
+        return book;
     }
 
     @DeleteMapping("{id}")
     public void deleteBook(@PathVariable String id) throws IOException {
         Book book = bookRepository.findById(id);
-        storageService.delete(book.getBook().getId().toString());
+        if (book.getBook() != null) {
+            storageService.delete(book.getBook().getId().toString());
+        }
+        
         bookRepository.deleteById(id);
     }
+
+    @PutMapping("{id}")
+    public Book editBook(
+        @PathVariable String id,
+        @RequestParam("model") String model,
+        @RequestParam(value = "file", required = false) MultipartFile file,
+        @RequestParam(value = "thumbnail", required = false) MultipartFile thumbnail
+    ) throws IOException {
+        ObjectMapper mapper = new ObjectMapper();
+        Book updateBook = mapper.readValue(model, Book.class);
+        Book book = bookRepository.findById(id);
+        
+        
+        // BeanUtils.copyProperties(updateBook, book);
+
+        book.setAuthor(updateBook.getAuthor());
+        book.setTitle(updateBook.getTitle());
+        book.setCategory(updateBook.getCategory());
+        book.setCollection(updateBook.getCollection());
+        book.setDescription(updateBook.getDescription());
+        book.setTags(updateBook.getTags());
+
+        // book.setFileType();
+        StoredFile tmpBook = storageService.load(book.getBookId());
+        if (file == null) {
+            book.setFileExtension(FilenameUtils.getExtension(tmpBook.getFilename()));
+            book.setBook(tmpBook);
+        }
+
+        if (thumbnail == null && book.getThumbnailId() != null) {
+            StoredFile tmpThumbnail = storageService.load(book.getThumbnailId());
+            book.setThumbnail(tmpThumbnail);
+        } else {
+            if (FilenameUtils.getExtension(tmpBook.getFilename()).equals("pdf")) {
+                book.storeThumbnail(thumbnail,  MediaType.APPLICATION_PDF);
+            } else {
+                book.storeThumbnail(thumbnail, MediaType.APPLICATION_OCTET_STREAM);
+            }
+            
+        }
+        
+        
+        bookRepository.save(book);
+        return book;
+    }
+
 
 
     @GetMapping("download/{id}")
@@ -148,13 +223,14 @@ public class BookController {
                 .body(new ByteArrayResource(loadFile.getFile()));
     }
 
+    Logger logger;
+
     @GetMapping("cbz-img/{id}/{filename}")
     public ResponseEntity<ByteArrayResource> getCbzFile(
         @PathVariable String id,
-        @PathVariable String filename
+        @PathVariable String filename,
+        @RequestParam(required = false) String progress
     ) throws IOException {
-        // final byte[] requestContent;
-        // requestContent = IOUtils.toByteArray(request.getReader());
         Book book = bookRepository.findById(id);
         StoredFile file = storageService.load(book.getBook().getId().toString());
         ZipInputStream zis = ZipService.unzipRef(new ByteArrayInputStream(file.getFile()));
@@ -162,6 +238,8 @@ public class BookController {
         // List<String> fileList = new ArrayList<String>();
         String fileNameDcd = new String(Base64.getDecoder().decode(filename));
         
+        
+        logger.info("Test");
         ZipEntry entry = zis.getNextEntry();
         // return "name.toString()";
         while(entry != null) {
@@ -174,7 +252,13 @@ public class BookController {
                     oStream.write(buffer, 0, len);
                 }
                 ResponseEntity.ok().contentType(MediaType.parseMediaType(MediaType.IMAGE_JPEG_VALUE));
-
+                oStream.close();
+                zis.close();
+                zis.closeEntry();
+                if (progress != null) {
+                    book.setProgress(progress);
+                    bookRepository.save(book);
+                }
                 return ResponseEntity.ok()
                         .contentType(MediaType.parseMediaType(MediaType.IMAGE_JPEG_VALUE ))
                         // .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + loadFile.getFilename() + "\"")
@@ -183,7 +267,8 @@ public class BookController {
             
             entry = zis.getNextEntry();
         }
-
+        zis.close();
+        zis.closeEntry();
         return null;
     }
 
@@ -200,6 +285,8 @@ public class BookController {
             
             entry = zis.getNextEntry();
         }
+        zis.closeEntry();
+        zis.close();
         Collections.sort(fileList);
         return fileList;
     }
